Align provider structure
