+++ src/features/image-generation.ts
