+++ src/vscode/proposed-chat-provider.d.ts
