+++ src/vscode/proposed-thinking-part.d.ts
